requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
CacheControl==0.12.10
cachetools==5.0.0
certifi==2021.10.8
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
Django==4.0.2
django-appconf==1.0.5
django-cors-headers==3.11.0
django-extensions==3.1.5
django-imagekit==4.1.0
django-jet==1.0.8
django-phonenumber-field==6.1.0
django-phonenumbers==1.0.1
django-rest-swagger==2.2.0
django-sslserver==0.22
djangorestframework==3.13.1
fcm-django==1.0.8
firebase-admin==5.2.0
google-api-core==2.5.0
google-api-python-client==2.37.0
google-auth==2.6.0
google-auth-httplib2==0.1.0
google-cloud-core==2.2.2
google-cloud-firestore==2.3.4
google-cloud-storage==2.1.0
google-crc32c==1.3.0
google-resumable-media==2.2.1
googleapis-common-protos==1.54.0
grpcio==1.44.0
grpcio-status==1.44.0
httplib2==0.20.4
idna==3.3
itypes==1.2.0
jet==1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
msgpack==1.0.3
openapi-codec==1.3.2
packaging==21.3
phonenumbers==8.12.44
pilkit==2.0
Pillow==9.0.1
proto-plus==1.20.3
protobuf==3.19.4
psycopg2==2.9.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.7
pytz==2021.3
requests==2.27.1
rsa==4.8
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
urllib3==1.26.8
